removed some classes we dont need. factored PID commands to a base class. fixed a couple of tests.
