+++ src/main/java/frc/robot/commands/AlignToAngleCommand.java
@@ -7,7 +7,6 @@ package frc.robot.commands;
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj.Joystick;
 import frc.robot.filters.DriveInput;
 import frc.robot.subsystems.DriveSubsystem;
@@ -17,20 +16,12 @@ import frc.robot.subsystems.DriveSubsystem;
  * 
  * @author aheitkamp
  */
-public class AlignToAngleCommand extends EntechCommandBase {
+public class AlignToAngleCommand extends BaseDrivePIDCommand {
 
-	public static final double P_GAIN = 0.0865;
-	public static final double I_GAIN = 0.5;
-    public static final double D_GAIN = 0.0075;
-    public static final double ANGLE_TOLERANCE = 1;
-    public static final double SPEED_LIMIT = 0.75;
-    public static final int STOP_COUNT = 4;
-    private final DriveSubsystem drive;
-    private final PIDController pid;
-    private final Joystick joystick;
     private Supplier<Double> yawAngleSupplier;
     private Supplier<Double> desiredAngleSupplier;
-    private StoppingCounter counter;
+
+    
     /**
      * Creates a new snap yaw degrees command that will snap the robot to the specified angle
      * 
@@ -40,18 +31,9 @@ public class AlignToAngleCommand extends EntechCommandBase {
      * @param joystick the joystick you controll the robot with
      */
     public AlignToAngleCommand(DriveSubsystem drive,  Joystick joystick,Supplier<Double> desiredAngleSupplier, Supplier<Double> yawAngleSupplier) {
-        super(drive);
-        this.drive = drive;
-        this.joystick = joystick;
-
-        pid = new PIDController(P_GAIN, I_GAIN, D_GAIN);
-        pid.enableContinuousInput(-180, 180);
-        pid.setTolerance(ANGLE_TOLERANCE);
-        counter = new StoppingCounter("AutoAlignCommand",STOP_COUNT);
-    }
-
-    @Override
-    public void initialize() {
+        super(drive,joystick);
+        this.desiredAngleSupplier = desiredAngleSupplier;
+        this.yawAngleSupplier = yawAngleSupplier;
     }
 
     @Override
@@ -75,18 +57,4 @@ public class AlignToAngleCommand extends EntechCommandBase {
         drive.drive(di );
     }
 
-    @Override
-    public void end(boolean interrupted) {
-        drive.brake();
-    }
-
-    @Override
-    public boolean isFinished() {
-    	return counter.isFinished(pid.atSetpoint());    	
-    }
-
-    @Override
-    public boolean runsWhenDisabled() {
-        return false;
-    }
 }
