+++ src/main/java/frc/robot/commands/AlignToScoringLocationCommand.java
@@ -3,7 +3,6 @@ package frc.robot.commands;
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj.Joystick;
 import frc.robot.filters.DriveInput;
@@ -20,20 +19,10 @@ import frc.robot.subsystems.DriveSubsystem;
  * 
  * @author dcowden
  */
-public class AlignToScoringLocationCommand extends EntechCommandBase {
+public class AlignToScoringLocationCommand extends BaseDrivePIDCommand {
 
-	public static final double P_GAIN = 0.0865;
-	public static final double I_GAIN = 0.5;
-    public static final double D_GAIN = 0.0075;
-    public static final double ANGLE_TOLERANCE = 1;
-    public static final double SPEED_LIMIT = 0.75;
-    public static final int STOP_COUNT = 4;
-    private final DriveSubsystem drive;
-    private final PIDController pid;
-    private final Joystick joystick;
-    ScoringLocation scoringLocation;
+    private ScoringLocation scoringLocation;
     private Supplier<Pose2d> currentPoseSupplier;
-    private StoppingCounter counter;
     private AlignmentCalculator alignCalculator = new AlignmentCalculator();
     
     
@@ -43,19 +32,9 @@ public class AlignToScoringLocationCommand extends EntechCommandBase {
      * @param joystick the joystick you controll the robot with
      */
     public AlignToScoringLocationCommand(DriveSubsystem drive,  Joystick joystick, ScoringLocation scoringLocation, Supplier<Pose2d> currentPoseSupplier) {
-        super(drive);
-        this.drive = drive;
-        this.joystick = joystick;
+        super(drive,joystick);
         this.scoringLocation = scoringLocation;
         this.currentPoseSupplier = currentPoseSupplier;
-        pid = new PIDController(P_GAIN, I_GAIN, D_GAIN);
-        pid.enableContinuousInput(-180, 180);
-        pid.setTolerance(ANGLE_TOLERANCE);
-        counter = new StoppingCounter("AlignToScoringLocationCommand",STOP_COUNT);
-    }
-
-    @Override
-    public void initialize() {
     }
 
     @Override
@@ -78,20 +57,5 @@ public class AlignToScoringLocationCommand extends EntechCommandBase {
         di.setOverrideYawLock(true);
         di.setOverrideAutoYaw(true);
         drive.drive(di );
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        drive.brake();
-    }
-
-    @Override
-    public boolean isFinished() {
-    	return counter.isFinished(pid.atSetpoint());    	
-    }
-
-    @Override
-    public boolean runsWhenDisabled() {
-        return false;
     }
 }
