--- src/main/java/frc/robot/pose/RobotCalculator.java
@@ -1,5 +1,0 @@
-package frc.robot.pose;
-
-public class RobotCalculator {
-
-}
