+++ src/main/java/frc/robot/subsystems/DriveSubsystem.java
@@ -12,13 +12,9 @@ import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj.drive.MecanumDrive;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-
 import frc.robot.RobotConstants;
-import frc.robot.filters.DriveFilterManager;
 import frc.robot.filters.DriveInput;
-import frc.robot.pose.AlignmentSolution;
 import frc.robot.pose.DriveStatus;
-import frc.robot.pose.RobotCalculations;
 
 /**
  *
