+++ src/main/java/frc/robot/subsystems/EntechSubsystem.java
@@ -8,22 +8,12 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.pose.RobotCalculations;
 
 public abstract class EntechSubsystem extends SubsystemBase {
 
     public EntechSubsystem() {
         CommandScheduler.getInstance().registerSubsystem(this);
     }
-    private RobotCalculations currentCalculations;
-    
-
-	public RobotCalculations getCurrentCalculations() {
-		return currentCalculations;
-	}
-	public void setCurrentCalculations(RobotCalculations currentCalculations) {
-		this.currentCalculations = currentCalculations;
-	}
 	public abstract void initialize();
     public abstract SubsystemStatus getStatus();
